feat: add basic testing mechanisms
This commit adds two features: first, the hice-dt executable will now
write its input files and output to a seperate directory. We do this
since they are otherwise overwritten with subsequent invocations.

Additionally, a script is added which can compare the output of two
different invocations and logs calls that are different.